moved params query
