+++ x/stakeibc/keeper/grpc_query_params.go
@@ -1,20 +1,1 @@
 package keeper
-
-import (
-	"context"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-
-	"github.com/Stride-Labs/stride/v18/x/stakeibc/types"
-)
-
-func (k Keeper) Params(c context.Context, req *types.QueryParamsRequest) (*types.QueryParamsResponse, error) {
-	if req == nil {
-		return nil, status.Error(codes.InvalidArgument, "invalid request")
-	}
-	ctx := sdk.UnwrapSDKContext(c)
-
-	return &types.QueryParamsResponse{Params: k.GetParams(ctx)}, nil
-}
